+++ src/patches/registry.cpp
@@ -3,10 +3,12 @@
 #include "knightCatapultCrash.h"
 #include "mpAiEnable.h"
 #include "mpConnectCompleteCrash.h"
+#include "endgameStats.h"
 
 void applyUnofficialPatches() {
     installKnightCatapultCrashFix();
     installMpAiEnable();
     installIntroSkip();
     installMpConnectCompleteCrashFix();
+    installEndgameStats();
 }